backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.5.1

# Analytics and Database Dependencies
sqlalchemy=2.0.23=pypi_0
psycopg2-binary=2.9.9=pypi_0
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.9.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
asttokens=3.0.0=pypi_0
bitsandbytes=0.45.0=pypi_0
blinker=1.9.0=py312h06a4308_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.7.15=h06a4308_0
certifi=2025.7.14=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.1.8=py312h06a4308_0
decorator=5.2.1=pypi_0
distro=1.9.0=pypi_0
executing=2.2.0=pypi_0
expat=2.7.1=h6a678d5_0
filelock=3.18.0=pypi_0
flask=3.1.0=py312h06a4308_0
flask_cors=3.0.10=pyhd3eb1b0_0
fsspec=2025.7.0=pypi_0
h11=0.16.0=pypi_0
hf-xet=1.1.5=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.34.0=pypi_0
idna=3.10=pypi_0
inflect=7.3.1=pypi_0
ipython=8.25.0=pypi_0
itsdangerous=2.2.0=py312h06a4308_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=py312h06a4308_0
jiter=0.10.0=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libmagic=5.46=h90f9352_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
lxml=5.3.0=pypi_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=3.0.2=py312h5eee18b_0
matplotlib-inline=0.1.7=pypi_0
more-itertools=10.7.0=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.5=pypi_0
numpy=2.3.2=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
openai=1.93.0=pypi_0
openssl=3.0.17=h5eee18b_0
packaging=25.0=pypi_0
parso=0.8.4=pypi_0
peft=0.14.1.dev0=pypi_0
pexpect=4.9.0=pypi_0
pillow=11.3.0=pypi_0
pip=25.1=pyhc872135_2
prompt-toolkit=3.0.51=pypi_0
psutil=7.0.0=pypi_0
pthread-stubs=0.3=h0ce48e5_1
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pyclamd=0.4.0=pypi_0
pydantic=2.11.7=pypi_0
pydantic-core=2.33.2=pypi_0
pygments=2.19.2=pypi_0
python=3.12.4=h5148396_1
python-dotenv=1.0.1=pypi_0
python-magic=0.4.27=pyh9ac5cc3_5
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.4=pypi_0
safetensors=0.5.3=pypi_0
setuptools=78.1.1=py312h06a4308_0
six=1.17.0=py312h06a4308_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.13.1=pypi_0
tk=8.6.14=h993c535_1
tokenizers=0.21.2=pypi_0
torch=2.5.1=pypi_0
torchaudio=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.49.0.dev0=pypi_0
triton=3.1.0=pypi_0
typeguard=4.4.4=pypi_0
typing-extensions=4.14.1=pypi_0
typing-inspection=0.4.1=pypi_0
tzdata=2025b=h04d1e81_0
urllib3=2.5.0=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.1.3=py312h06a4308_0
wheel=0.45.1=py312h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
gunicorn=23.0.0=pypi_0
